abena_biomaker_integration/Abena Biomarker Integration/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
websockets==11.0.3
requests>=2.26.0
aiohttp==3.8.5
pyserial>=3.5
bleak>=0.19.0
matplotlib==3.7.2
plotly==5.16.1
dash==2.13.0
dash-bootstrap-components==1.5.0
flask==2.2.5
python-dotenv>=0.19.0
PyJWT>=2.3.0 